Melhorias na interface
Falta arrumar as mensagens de erro da tela login e quando elas aparecem (talvez usando um contador estatíco)

+++ views/listagemCategoria.php
@@ -16,21 +16,25 @@
 
 <div class='container-categorias'>
     <?php
-        foreach($categorias as $categoria){
-            $caminho = APP;
-            echo "
-                <div class='categoria'>
-                    <p class='nome-categoria'>{$categoria['descricao']}</p>
-                    <div class='container-buttons'>
-                        <a class='butao editar-categoria' type='button' href='$caminho/categoria/editar/{$categoria['id']}'>
-                            <p class='buton-icon'>*</p>
-                        </a>
-                        <a class='butao deletar-categoria' type='button' href='$caminho/categoria/excluir/{$categoria['id']}/{$usuarioLogado['id']}'>
-                            <p class='buton-icon'>X</p>
-                        </a>
+        if(!empty($categorias)){
+            foreach($categorias as $categoria){
+                $caminho = APP;
+                echo "
+                    <div class='categoria'>
+                        <p class='nome-categoria'>{$categoria['descricao']}</p>
+                        <div class='container-buttons'>
+                            <a class='butao editar-categoria' type='button' href='$caminho/categoria/editar/{$categoria['id']}'>
+                                <p class='buton-icon'>*</p>
+                            </a>
+                            <a class='butao deletar-categoria' type='button' href='$caminho/categoria/excluir/{$categoria['id']}/{$usuarioLogado['id']}'>
+                                <p class='buton-icon'>X</p>
+                            </a>
+                        </div>
                     </div>
-                </div>
-            ";
+                ";
+            }
+        } else{
+            echo "<p class='msg-nenhuma'>Nenhuma categoria cadastrada!</p>";
         }
     ?>
 </div>
